fix: drop the int cast in the OmegaIm default (0.01 * LargeValue)
The spectrum default was computed as 0.01 * (int)StdI->LargeValue,
truncating the fractional part of LargeValue: for LargeValue < 1 the
default collapsed to exactly 0, and for any non-integer LargeValue the
intended 0.01 * LargeValue was rounded down to the previous 0.01 step.

Survey over the reference suite (65 modpara.def with OmegaIm): 34 cases
used the truncated default and change; 24 have integer LargeValue
(unchanged); 7 set OmegaIm explicitly (unchanged). The reference files
are updated surgically - only the OmegaIm column of the OmegaMax /
OmegaMin lines changes; base-mode integration passes 94/94.

Behavior change: HPhi spectrum / time-evolution runs that relied on the
truncated default will see a slightly larger OmegaIm (or a non-zero one
where it used to be 0).

+++ src/StdFace_main.c
@@ -1970,7 +1970,7 @@ static void CheckModPara(struct StdIntList *StdI)
   StdFace_PrintVal_d("OmegaMax", &StdI->OmegaMax, StdI->LargeValue*StdI->nsite);
   StdFace_PrintVal_d("OmegaMin", &StdI->OmegaMin, -StdI->LargeValue*StdI->nsite);
   StdFace_PrintVal_d("OmegaOrg", &StdI->OmegaOrg, 0.0);
-  StdFace_PrintVal_d("OmegaIm", &StdI->OmegaIm, 0.01* (int)StdI->LargeValue);
+  StdFace_PrintVal_d("OmegaIm", &StdI->OmegaIm, 0.01 * StdI->LargeValue);
 #elif defined(_mVMC)
   if (strcmp(StdI->CParaFileHead, "****") == 0) {
     strcpy(StdI->CParaFileHead, "zqp\0");
